% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepwalker_verse.R
\name{deepwalker_verse}
\alias{deepwalker_verse}
\title{Versatile Graph Embeddings from Similarity Measures}
\usage{
deepwalker_verse(x, similarity = c("pagerank", "adjacency", "simrank"),
  dimension = 128, epochs = 1e+05, learning_rate = 0.0025, samples = 3,
  alpha = 0.85)
}
\arguments{
\item{x}{an object of class dgRMatrix}

\item{similarity}{type of similarity. Currently defaults to pagerank.}

\item{dimension}{integer with the dimension of the embedding}

\item{epochs}{number of epochs}

\item{learning_rate}{the learning rate of the algorithm}

\item{samples}{integer}

\item{alpha}{the alpha parameter of the algorithm}
}
\value{
a list with elements 
\enumerate{
\item{vertices: }{the number of vertices in the graph}
\item{edges: }{the number of edges (links) between the vertices in the graph}
\item{embedding: }{the embeddings of the vertices}
\item{edges_data and offsets_data: }{for debugging purposes now}
}
}
\description{
This function calculates node embeddings in a graph by using 
VERSE (Versatile Graph Embeddings from Similarity Measures)
}
\examples{
library(igraphdata)
library(igraph)
library(Matrix)

data(karate, package = "igraphdata")
x <- as_adj(karate)
x <- as(x, "RsparseMatrix")

node_embeddings <- deepwalker_verse(x, dimension = 64, epochs = 1000,
    learning_rate = 0.0025, samples = 3, alpha = 0.85)
str(node_embeddings$embeddings)
}
