% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepwalker_verse.R
\name{deepwalker_verse}
\alias{deepwalker_verse}
\title{Versatile Graph Embeddings from Similarity Measures}
\usage{
deepwalker_verse(x)
}
\arguments{
\item{x}{an object of class dgRMatrix}
}
\value{
TODO
}
\description{
This function calculates node embeddings in a graph by using 
VERSE (Versatile Graph Embeddings from Similarity Measures)
}
\examples{
library(igraphdata)
library(igraph)
library(Matrix)

data(karate, package = "igraphdata")
x <- as_adj(karate)
x <- as(x, "RsparseMatrix")

embeddings <- deepwalker_verse(x)
}
